# openapi.json
{"openapi":"3.1.0","info":{"title":"Revenue Recovery Agent","version":"0.1.0"},"paths":{"/webhooks/razorpay":{"post":{"tags":["webhooks"],"summary":"Razorpay Webhook","operationId":"razorpay_webhook_webhooks_razorpay_post","parameters":[{"name":"x-razorpay-signature","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Razorpay-Signature"}},{"name":"x-razorpay-event-id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Razorpay-Event-Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Razorpay Webhook Webhooks Razorpay Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/recovery/metrics":{"get":{"tags":["recovery"],"summary":"Recovery Metrics","operationId":"recovery_metrics_recovery_metrics_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/recovery/metrics/breakdowns":{"get":{"tags":["recovery"],"summary":"Recovery Metric Breakdowns","operationId":"recovery_metric_breakdowns_recovery_metrics_breakdowns_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/recovery/cases/{case_id}/timeline":{"get":{"tags":["recovery"],"summary":"Recovery Case Timeline","operationId":"recovery_case_timeline_recovery_cases__case_id__timeline_get","parameters":[{"name":"case_id","in":"path","required":true,"schema":{"type":"string","title":"Case Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/recovery/cases/{case_id}/escalation":{"get":{"tags":["recovery"],"summary":"Recovery Case Escalation","operationId":"recovery_case_escalation_recovery_cases__case_id__escalation_get","parameters":[{"name":"case_id","in":"path","required":true,"schema":{"type":"string","title":"Case Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/recovery/cases/{case_id}/escalation/assignment":{"patch":{"tags":["recovery"],"summary":"Assign Recovery Case Escalation","operationId":"assign_recovery_case_escalation_recovery_cases__case_id__escalation_assignment_patch","parameters":[{"name":"case_id","in":"path","required":true,"schema":{"type":"string","title":"Case Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EscalationAssignmentRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/recovery/cases/{case_id}/escalation/notes":{"post":{"tags":["recovery"],"summary":"Add Recovery Case Escalation Note","operationId":"add_recovery_case_escalation_note_recovery_cases__case_id__escalation_notes_post","parameters":[{"name":"case_id","in":"path","required":true,"schema":{"type":"string","title":"Case Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EscalationNoteRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/recovery/cases/{case_id}/escalation/resolve":{"post":{"tags":["recovery"],"summary":"Resolve Recovery Case Escalation","operationId":"resolve_recovery_case_escalation_recovery_cases__case_id__escalation_resolve_post","parameters":[{"name":"case_id","in":"path","required":true,"schema":{"type":"string","title":"Case Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Check Health Get"}}}}}}}},"components":{"schemas":{"EscalationAssignmentRequest":{"properties":{"assigned_team":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Assigned Team"},"assigned_to":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Assigned To"}},"type":"object","title":"EscalationAssignmentRequest"},"EscalationNoteRequest":{"properties":{"note":{"type":"string","title":"Note"},"created_by":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Created By"}},"type":"object","required":["note"],"title":"EscalationNoteRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}